db/db_config.py
import mysql.connector

def get_connection():

    conn = mysql.connector.connect(
        host="localhost",
        user="root",
        password="root",
        database="spydx_productpulse"
    )
    return conn
